[WordSeg] Move dataset into Datasets
Rename `DataSet` to `WordSegDataset` and move to Datasets/WordSeg.
Move Vocabularies.swift into ModelSupport.
Split Vocabularies.swift into 3 separate files, one for each contained struct: Alphabet.swift, CharacterSequence.swift, and Lexicon.swift.
