+++ Models/Text/CMakeLists.txt
@@ -13,12 +13,10 @@ add_library(TextModels
   TransformerBERT.swift
   Utilities.swift
   WeightDecayedAdam.swift
-  WordSeg/DataSet.swift
-  WordSeg/Model.swift
   WordSeg/Lattice.swift
+  WordSeg/Model.swift
   WordSeg/SE-0259.swift
-  WordSeg/SemiRing.swift
-  WordSeg/Vocabularies.swift)
+  WordSeg/SemiRing.swift)
 set_target_properties(TextModels PROPERTIES
   INTERFACE_INCLUDE_DIRECTORIES ${CMAKE_Swift_MODULE_DIRECTORY})
 target_compile_options(TextModels PRIVATE
