+++ Tests/TextTests/XCTestManifests.swift
@@ -18,7 +18,7 @@ import XCTest
   public func allTests() -> [XCTestCaseEntry] {
     return [
       testCase(TextInferenceTests.allTests),
-      testCase(WordSegDataSetTests.allTests),
+      testCase(WordSegDatasetTests.allTests),
       testCase(WordSegSemiRingTests.allTests),
       testCase(WordSegVocabularyTests.allTests),
       testCase(WordSegProbeLayerTests.allTests),
